Dados da API na base e retorno DataDevice na API

+++ Core/ApiModel/Request/DeviceGlobalRequest.cs
@@ -81,7 +81,7 @@ namespace Core.ApiModel.Request
         public double RmsAfSpdY { get; set; }
 
         [JsonProperty("rms_af_spd_Z")]
-        public int RmsAfSpdZ { get; set; }
+        public double RmsAfSpdZ { get; set; }
     }
 
     public class SetupAf
